adds source_coordinate and source_coordinate_range (closes #1)
The compiler front-end will need a way to indicate where source mappings
are located in a source file. This can be achieved using coordinates to
specific points in a source file, and coordinate ranges.

A source coordinate weakly models a constant iterator, and a source
coordinate range weakly models a range of characters between two source
coordinates.