implements the diagnostic level and diagnostic base types
Closes #3

From the design specification:

>    All diagnostics should store a range of coordinates to source
>    tokens, and a help message. There will be three diagnostic types:
>    ill-formed, warning, and remark.

Since the underlying implementation of one diagnostic is the same as
another, it makes sense to have a common base for every diagnostic, to:

1. implement [DRY]
2. take advantage of [EBO]

The base class will rely on a diagnostic level type, so it makes sense
to couple it in #3.

Diagnostics are intended to be privately derived from the diagnostic
base, although the implementation has no way to make this guarantee.

[DRY]: https://en.wikipedia.org/wiki/Don%27t_repeat_yourself
[EBO]: https://en.cppreference.com/w/cpp/language/ebo