enables all is_escape.cpp checks (closes #25)
The checks for exhaustive_check_is_unicode_escape were only being run
for N == 2 when using a function. A bit of investigating revealed that
the test would only run the checks for the _first_ occurrence of
exhaustive_check_is_unicode_escape, regardless of N's size.

This problem evaporated for unknown reasons when using a macro.